requirements.txt
streamlit
numpy
faiss-cpu
sentence-transformers
PyMuPDF
python-docx
requests
langchain